Specify localhost target inside setup-ansible-itself
It doesn't make sense to use inventory for this playbook.  The only
usecase is runnning it on localhost, so we may as well specify it insde
the playbook itself.

Note that we could potentially want to manage ansible configuration on
multiple machines, but then we would have to find a way for supplying
sudo passwords for all these machines (which is doable but adds
complexity).