Define all (?) locations in group_vars
This is tricky, as these have to be configured in both dotfiles
repository and here. As far as I can see, the original source of truth
about these variables must live in ansible repo (because it appears
earlier in configuration lfecycle - dotfiles are cloned *using*
ansible). I think I will template a file with bash env variables with an
ansible task, and that will be the source of this information in
dotfiles themselves.